feat: refactor Configuration to live in a Scope
The Configuration now lives inside a scope which will be passed to all objects instead of the direct Configuration. This makes it possible to change the configuration and reload all components depending on it in a ordered manner.

BREAKING CHANGE: every art class now takes a scope instead of a configuration. The scope is now the main entry point into the art-framework.

Signed-off-by: Michael Reichenbach <michael@reichenbach.in>